requirements.txt
backports.zoneinfo==0.2.1
importlib-metadata==7.0.0
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
python-frontmatter==1.0.1
PyYAML==5.1
tzlocal==5.2
zipp==3.17.0
